Changed gifts service entry point to TypeScript
no issue

The gifts module needed a typed entry point so TypeScript consumers can use its public interface without declaration shims or error suppressions.
