--- ghost/core/core/server/services/gifts/index.js
@@ -1,3 +1,0 @@
-const GiftServiceWrapper = require('./gift-service-wrapper');
-
-module.exports = new GiftServiceWrapper();
